+++ buildSrc/src/main/kotlin/BuildConfig.kt
@@ -3,10 +3,10 @@ package com.unciv.build
 object BuildConfig {
     const val kotlinVersion = "1.3.71"
     const val appName = "Unciv"
-    const val appCodeNumber = 488
-    const val appVersion = "3.11.5"
+    const val appCodeNumber = 494
+    const val appVersion = "3.11.9-patch2"
 
-    const val gdxVersion = "1.9.10"
+    const val gdxVersion = "1.9.12"
     const val roboVMVersion = "2.3.1"
     const val ashleyVersion = "1.7.0"
     const val aiVersion = "1.8.0"