+++ core/src/com/unciv/logic/battle/Battle.kt
@@ -63,7 +63,7 @@ object Battle {
         postBattleNationUniques(defender, attackedTile, attacker)
 
         // This needs to come BEFORE the move-to-tile, because if we haven't conquered it we can't move there =)
-        if (defender.isDefeated() && defender is CityCombatant && attacker.isMelee())
+        if (defender.isDefeated() && defender is CityCombatant && attacker is MapUnitCombatant && attacker.isMelee() && !attacker.unit.hasUnique("Unable to capture cities"))
             conquerCity(defender.city, attacker)
 
         // we're a melee unit and we destroyed\captured an enemy unit
@@ -122,9 +122,7 @@ object Battle {
         }
 
         for (unique in bonusUniques) {
-            if (!defeatedUnit.matchesCategory(unique.params[1])) {
-                continue
-            }
+            if (!defeatedUnit.matchesCategory(unique.params[1])) continue
 
             val yieldPercent = unique.params[0].toFloat() / 100
             val defeatedUnitYieldSourceType = unique.params[2]
@@ -303,13 +301,13 @@ object Battle {
         attackerCiv.addNotification("We have conquered the city of [${city.name}]!", city.location, Color.RED)
 
         city.getCenterTile().apply {
-            if(militaryUnit!=null) militaryUnit!!.destroy()
-            if(civilianUnit!=null) captureCivilianUnit(attacker, MapUnitCombatant(civilianUnit!!))
-            for(airUnit in airUnits.toList()) airUnit.destroy()
+            if (militaryUnit != null) militaryUnit!!.destroy()
+            if (civilianUnit != null) captureCivilianUnit(attacker, MapUnitCombatant(civilianUnit!!))
+            for (airUnit in airUnits.toList()) airUnit.destroy()
         }
         city.hasJustBeenConquered = true
 
-        if (!attackerCiv.isMajorCiv()){
+        if (!attackerCiv.isMajorCiv()) {
             city.destroyCity()
             return
         }
@@ -317,10 +315,9 @@ object Battle {
         if (attackerCiv.isPlayerCivilization()) {
             attackerCiv.popupAlerts.add(PopupAlert(AlertType.CityConquered, city.id))
             UncivGame.Current.settings.addCompletedTutorialTask("Conquer a city")
-        }
-        else {
+        } else {
             city.puppetCity(attackerCiv)
-            if (city.population.population < 4) {
+            if (city.population.population < 4 && !city.isOriginalCapital) {
                 city.annexCity()
                 city.isBeingRazed = true
             }
@@ -328,9 +325,9 @@ object Battle {
     }
 
     fun getMapCombatantOfTile(tile:TileInfo): ICombatant? {
-        if(tile.isCityCenter()) return CityCombatant(tile.getCity()!!)
-        if(tile.militaryUnit!=null) return MapUnitCombatant(tile.militaryUnit!!)
-        if(tile.civilianUnit!=null) return MapUnitCombatant(tile.civilianUnit!!)
+        if (tile.isCityCenter()) return CityCombatant(tile.getCity()!!)
+        if (tile.militaryUnit != null) return MapUnitCombatant(tile.militaryUnit!!)
+        if (tile.civilianUnit != null) return MapUnitCombatant(tile.civilianUnit!!)
         return null
     }