+++ desktop/src/com/unciv/app/desktop/DesktopLauncher.kt
@@ -25,7 +25,6 @@ import io.ktor.server.netty.Netty
 import java.io.File
 import java.util.*
 import kotlin.concurrent.timer
-import kotlin.system.exitProcess
 
 internal object DesktopLauncher {
     private var discordTimer: Timer? = null
@@ -48,7 +47,6 @@ internal object DesktopLauncher {
 
         val desktopParameters = UncivGameParameters(
                 versionFromJar,
-                exitEvent = { exitProcess(0) },
                 cancelDiscordEvent = { discordTimer?.cancel() },
                 fontImplementation = NativeFontDesktop(ORIGINAL_FONT_SIZE.toInt()),
                 customSaveLocationHelper = CustomSaveLocationHelperDesktop()
@@ -125,8 +123,10 @@ internal object DesktopLauncher {
          * this causes a delay, leading to horrible lag if there are enough switches.
          * The cost of this specific solution is that the entire game.png needs be be kept in-memory constantly.
          * It's currently only 1.5MB so it should be okay, but it' an important point to remember for the future.
-         * Sidenode: Modded tilesets don't have this problem, since all the images are included in the mod's single PNG.
+         * Sidenote: Modded tilesets don't have this problem, since all the images are included in the mod's single PNG.
          * Probably. Unless they have some truly huge images.
+         * Sidenote 2: Okay entire so custom tilesets do have this problem because they can get so big that what accommodates
+         * the regular tileset (4096) doesn't suit them. Looking at you 5hex.
          */
         settings.maxWidth = 4096
         settings.maxHeight = 4096